add: Vue Routerの設定まで #3
