+++ app/javascript/packs/router/router.js
@@ -1,0 +1,11 @@
+import Vue from 'vue'
+import VueRouter from 'vue-router'
+import BookHome from '../BookHome.vue'
+
+Vue.use(VueRouter)
+
+const routes = [
+  { path: '/', name: 'BookHome', component: BookHome }
+];
+
+export default new VueRouter({ routes });
