+++ package.json
@@ -6,6 +6,7 @@
     "axios": "^0.21.1",
     "vue": "^2.6.12",
     "vue-loader": "^15.9.6",
+    "vue-router": "^3.5.1",
     "vue-template-compiler": "^2.6.12"
   },
   "devDependencies": {
